pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.0.2"
description = "MLOps platform using MLflow (and some flyte)"
authors = ["Stefano Clarence"]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
python-dotenv = "^1.0.1"

[tool.poetry.group.flyte.dependencies]
flytekit = "1.10.3"
flytekitplugins-pod = "1.10.3"


[tool.poetry.group.mlflow.dependencies]
mlflow = "2.10.2"
aiobotocore = "2.5.4"
s3fs = "2023.9.2"
boto3 = "1.28.17"
botocore = "1.31.17"
fsspec = "2023.9.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[tool.poetry.group.inference.dependencies]
pycaret = "3.2.0"
kaleido = "0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
